drone/requirements.txt
streamlit
requests
pillow
opencv-python
pip install requirements.txt